// src/types.d.ts
// Define the Book interface
export interface Book {
  title: string;
  pages: number;
  genre: string;
  cover: string;
  synopsis: string;
  year: number;
  ISBN: string;
  author: {
    name: string;
    otherBooks: string[];
  };
}

// Define the Library interface
export interface Library {
  book: Book;
}